Initial commit
This is the last version I had lying around prior to the deletion. Sadly
my copy in neovim seems to be shallow or whatever so I do not have a
the full history. I also was having trouble pushing the "grafted" tagged
commits as a fake history, sooooo I am just making a clean commit of
this version I have. No history left here for now.

The commit in question that it was based on was
9adf692d7ae6424038458a89d4a411f0a27d1388 from ackyshake/VimCompletesMe
committed on 17 or 18 February 2022. The repository seems to have
disappeared from Github, where it was hosted. Web archive has this:
https://web.archive.org/web/20221126214126/https://github.com/ackyshake/VimCompletesMe.git
but alas the Git repository is not something they seem to archive
(unless I missed it).

Anyway, it's released under the Vim license, so I can dump a copy here.